Guard against concurrent modification of whitelistEntry
Modifying blacklistEntry is thread-safe as it uses a thread-safe data 
structure (CopyOnWriteArray). whitelistEntry is not, as while it uses
a concurrent data structure to track the list of patterns, several
threads / HTTP calls may attempt to set the whitelist or clear the
whitelist concurrently.

Protecting these calls by synchronizing on the BrowserMobHttpClient 
instance should protect againt concurrent modification of the list, and
associated race conditions.