New method for getting a connectable proxy address
The original getLocalHost() method was fine for returning an address to bind to,
but for the times when you wanted to find out what address to connect to (e.g.
when constructing a proxy address in seleniumProxy()), returning 0.0.0.0 would
cause problems.

This is an additional method that will filter out the uncontactable 0.0.0.0 address,
and replace it with a more contactable localhost address. seleniumProxy() method
updated to use the new method.