Fixes HarResultsTest on Windows
The headers coming back from the server depend on which platform Jetty
is running on. From evidence on github and my own machine, it seems that
they're 226 bytes long on Win, and 227 elsewhere.

That being said, the sample size is very low, and it would be better to
calculate the expected header size as part of the test. But I'm not sure
I understand the intention behind the test well enough to do that.