Add note about using release profile to create bin
I don't know too much about Java or Maven, and it took me a while to figure out how to create the binaries that are precompiled and included in the zip download. It's probably second nature for y'all, but this line would've saved me a few hours of headache. What do y'all think?